// apps/backend/src/common/rbac/permission.enum.ts
export enum Permission {
  RECEIVABLE_READ = 'RECEIVABLE_READ',
  RECEIVABLE_WRITE = 'RECEIVABLE_WRITE',
  RECEIVABLE_WRITE_OFF = 'RECEIVABLE_WRITE_OFF',
  RECEIVABLE_DISPUTE = 'RECEIVABLE_DISPUTE',
  RECEIVABLE_IMPORT = 'RECEIVABLE_IMPORT',
  PAYMENT_ALLOCATE = 'PAYMENT_ALLOCATE',
  PAYMENT_ALLOCATE_UNDO = 'PAYMENT_ALLOCATE_UNDO',
  EMAIL_TEMPLATE_READ = 'EMAIL_TEMPLATE_READ',
  REMINDER_POLICY_WRITE = 'REMINDER_POLICY_WRITE',
  REMINDER_SEND_MANUAL = 'REMINDER_SEND_MANUAL',
  BANK_CONNECTION_READ = 'BANK_CONNECTION_READ',
  BANK_CONNECTION_MANAGE = 'BANK_CONNECTION_MANAGE',
  SUBSCRIPTION_MANAGE = 'SUBSCRIPTION_MANAGE',
  USER_MANAGE = 'USER_MANAGE',
  INTERNAL_TASK_MANAGE = 'INTERNAL_TASK_MANAGE',
  REPORT_READ = 'REPORT_READ',
  AUDIT_LOG_READ = 'AUDIT_LOG_READ',
  SWITCH_ORGANIZATION = 'SWITCH_ORGANIZATION',
  CUSTOMER_READ = 'CUSTOMER_READ',
  ORGANIZATION_READ = 'ORGANIZATION_READ',
}